+++ Applications/MIDIMonitor/SMMSysExWindowController.m
@@ -55,7 +55,7 @@ NSString* const SMMSaveSysExWithEOXAlwaysPreferenceKey = @"SMMSaveSysExWithEOXAl
 
 #pragma mark Private
 
-- (void)savePanelDidEnd:(NSSavePanel *)sheet returnCode:(int)returnCode contextInfo:(void *)contextInfo;
+- (void)savePanelDidEnd:(NSSavePanel *)sheet returnCode:(NSInteger)returnCode contextInfo:(void *)contextInfo;
 {
     [sheet orderOut:nil];
