+++ Applications/SysExLibrarian/SSEFindMissingController.m
@@ -21,10 +21,10 @@
 @interface SSEFindMissingController (Private)
 
 - (void)findNextMissingFile;
-- (void)missingFileAlertDidEnd:(NSWindow *)sheet returnCode:(int)returnCode contextInfo:(void *)contextInfo;
+- (void)missingFileAlertDidEnd:(NSWindow *)sheet returnCode:(NSInteger)returnCode contextInfo:(void *)contextInfo;
 
 - (void)runOpenSheetForMissingFile;
-- (void)findMissingFileOpenPanelDidEnd:(NSOpenPanel *)openPanel returnCode:(int)returnCode contextInfo:(void *)contextInfo;
+- (void)findMissingFileOpenPanelDidEnd:(NSOpenPanel *)openPanel returnCode:(NSInteger)returnCode contextInfo:(void *)contextInfo;
 
 @end
 
@@ -92,7 +92,7 @@
     }
 }
 
-- (void)missingFileAlertDidEnd:(NSWindow *)sheet returnCode:(int)returnCode contextInfo:(void *)contextInfo;
+- (void)missingFileAlertDidEnd:(NSWindow *)sheet returnCode:(NSInteger)returnCode contextInfo:(void *)contextInfo;
 {
     if (returnCode == NSAlertDefaultReturn) {
         // Get this sheet out of the way before we open another one
@@ -117,7 +117,7 @@
     }];
 }
 
-- (void)findMissingFileOpenPanelDidEnd:(NSOpenPanel *)openPanel returnCode:(int)returnCode contextInfo:(void *)contextInfo;
+- (void)findMissingFileOpenPanelDidEnd:(NSOpenPanel *)openPanel returnCode:(NSInteger)returnCode contextInfo:(void *)contextInfo;
 {
     BOOL cancelled = NO;
